+++ AdminAction.java
@@ -2,5 +2,5 @@ package mymessenger;
 
 public interface AdminAction extends ModeratorAction{
     public void showAllUsers();
-    public void deleteUser(Long id);
+    public void deleteUser(Integer id);
 }
