+++ UserAction.java
@@ -1,5 +1,5 @@
 package mymessenger;
 
 public interface UserAction extends BaseUserAction<BaseUser, String>{
-    public void editMessage(Long id, String newText);
+    public void editMessage(Integer id, String newText);
 }
